Move anomaly lightning model
