--- src/otx/core/model/module/anomaly/__init__.py
@@ -1,8 +1,0 @@
-"""Anomaly models."""
-
-# Copyright (C) 2024 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-
-from .anomaly_lightning import OTXAnomaly
-
-__all__ = ["OTXAnomaly"]
