+++ src/otx/core/model/module/anomaly_lightning.py
